requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
asyncio==3.4.3
aiohttp==3.9.1
httpx==0.25.2

# API Clients
requests==2.31.0
gql==3.5.0
websockets==12.0

# Blockchain / Web3
web3==6.11.3
eth-account==0.10.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Scheduling
apscheduler==3.10.4

# Monitoring & Alerts
python-telegram-bot==20.7
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.4

# API Framework
fastapi==0.108.0
uvicorn==0.25.0

# Security
cryptography==41.0.7
boto3==1.34.10

# Utilities
colorlog==6.8.0
click==8.1.7
tabulate==0.9.0
jsonschema==4.20.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==21.0.0

# Development
black==23.12.1
flake8==6.1.0
mypy==1.8.0
